Fix broken input on OnTriggerStay2D methods
For some reason, when getting button presses in OnTriggerStay, not all
button presses are caught. This commit fixes this behavior by using
OnTriggerStay2D to set internal state, and reading input in the
Update method.

This commit also organizes the order of components in the screen.
Background is 0, mechanics are 1, player is 2, walls and platforms are
3.